update powertip to lastest dev version 0cf96e91295
This version includes this bugfix:
https://github.com/stevenbenner/jquery-powertip/issues/79#issuecomment-20772860

Powertip got the wrong placenment when moving between at least two
friends, the first powertip did not have time to to close before showing
the next. This would fail to keep the placement to 'nw' and using auto
instead, because it would remove that class after closeDelay. Setting
closeDelay to 0 also 'fixed' it, but was to hard to hover the popup.
http://i.imgur.com/8HR1EZJ.png