rename Nightly tournaments to Oriental
Daily tournaments are scheduled on lichess peak activity hours,
which match the Western world day.
Therefore nightlies are targetting the Oriental world.

- Why not call them nightlies?
because it's not night for those who play them

- Why not call them dailies?
because they won't be very successful, and I don't want
a half-empty tournament to be called "daily".
It would look bad to have weak daily tournaments.

This is a proposition, please discuss.