crazy hack around rematch bug
the idea is that when a rematch is accepted,
the browser does two things:
1. redirect to the new game
2. reload current game data with XHR
the XHR call may fail depending on the browser behavior during redirect.
a failure to reload game data through XHR triggers a full page reload.
If the game was being played with a spectator URL,
the reload brings the player to the analysis page.

This is all very conjectural