General improvements to the README file.
Edited some of the wording so that it would read better. Also removed
commas when commas are not needed. There were a few commas before the
word 'and' when the word 'and' was not separating to independent clauses..

Added a License subheading at the bottom. The README didn't
mention which license is used. It should be mentioned in the README so
others don't have to click the License text file just to know what the
license is.

Other minor edits regarding wordy sentence and what not.