Avoid links to ongoing private tournaments (fixes #1691)
When people are in a private tournament, a link to the tournament is still available on their profile, or on the sidebar of an ongoing tournament game, or using the "View tournament" button on TV. This commit makes sure that those links are not displayed for ongoing tournaments. When the tournaments are finished, the links can be displayed because joining is no longer possible, so there's no reason to keep those tournaments private.