Fix a bug for study icon in friend list
StudyBecamePublic/StudyBecamePrivate/... did not take the studyId in
account when comparing against onlineStudyingAll/onlineStudying. This
caused bugs like this one: if you are working in a private study, and you
are also a contributor of another private study (but not working there at
the moment), and this latter study becomes public, then the icon would
show up when it should not.

To make this happen, 7e2ec98... had to be reverted.