Don't linkify expanded images.
Links suggest to the user that additional information is available
by clicking through. But on expanded images, the link goes to the
exact same content as the inline image. So instead the user has to
remember that image links aren't going somewhere helpful.

Users can still right click -> view image to get to the image.