Make losses to AI show up under 'Losses'
Fixes #3815.

To remove the AI games from the 'Draws' tab, we can use the winnerColor field instead of the winnerId field - the latter does not exist on AI games, but the former does. So only games without winnerColor are definitely drawn.

To make them show up under losses, we use the fact that games must be lost if winnerColor exists but winnerId not - this only happens if you lose against the AI (if you draw against the ID, winnerColor won't exist; if you win against the ID or play against humans, winnerId will exist).