Rating graph on perfStat page, closes #2425
On each perfStat page, a rating graph is drawn for the specific perf (in the same color as its line on the complete profile rating graph). If there is no data for a certain perf, `lichess.ratingHistoryChart` will take care of hiding the assigned `<div>` as to not have a big white space on top of the perfStat page.

`ratingHistoryChart` also takes care of filtering out the non-perf lines -- this cannot happen on the server side because the rating graph cache contains all graphs.