Store history events in ArrayDeque
Switch underlying data structure from List to java ArrayDeque,
which allows fast insert and deletion from head. Store events
in chronological order, consistent with the order on disk.

This change allows us to remove a bunch of reverse() calls. Once
we upgrade to scala 2.13 we'll be able to migrate to scala's
ArrayDeque and remove the java <-> scala conversions.