Fix Quit data race
Improve our reconnect race bandaid. The first version
of the bandaid uses a boolean flag which might not have been
set yet, because the new Socket might be in a message queue
instead of fully initiated. With the first version, we could
end up queueing a Quit message that gets processed after
the socket is reconnected.

This version does the check on the actor thread.