Keep track of what positions people tend to sit or disconnect in
Not used for anything else than mod zone display at the moment, if the value seems usable then it can be used in the punishment code.

The counter gets +1'd if you disconnect or sit in a winning position (material imbalance >= 5) and -1 of you disconnect or sit in a losing position (material imbalance <= -5).
The idea is that abusers will get a highly negative score and people with an unstable connection will have a score around zero.