Resume Howler's AudioContext when suspended
On Chrome, it sometimes (and irregularly) happens that the AudioContext is not allowed to start on the first sound. It has to be explicitly resumed after a "user gesture" (such as a click) on the page.

If we don't have any resume() calls, then a silenced Chrome page would just stay silenced. Attempting to resume whenever the AudioContext is suspended will ensure that a blocked AudioContext can still start after a user clicks on the page.