Fix auto-reloading of mod/public-chat
It has been broken for a while, since the $("#lichess") selector started returning null.

The #comm-wrap div is needed because not introducing a new element but using an existing one as replacement for $("#lichess") always resulted in a wrong layout after an auto-refresh.

I'm seeing a Content Security Policy error every time .load happens. I don't think this is a big deal (everything works), but just mentioning it in case.