Use prefers-color-scheme if no explicit pref is set
If an explicit preference for the background color has been set, lila will then use that one always. If such a preference is not set, the prefers-color-scheme media query is used - if the user prefers a dark scheme then lila loads the dark CSS, if the value of the media query is 'light' or 'no-preference' then the light CSS is loaded.

The server gets to know about the prefers-color-scheme value through a prefers-dark cookie that gets set to 1 if (prefers-color-scheme: dark) is true. On the first request, that cookie does not yet exist, and then site.js may have to reload the page if setting the cookie value is likely to change the background.