Hourly Atomic and Antichess tournaments
Antichess and atomic now have as many players per week as crazyhouse did when the hourly crazyhouse arenas were added. This will also make it so there is always an antichess and atomic tournament on the /tournaments page, making it easier for users to find an ongoing antichess/atomic tournament.