Remove "reportable && exceptionalDif" condition
I didn't know that this happened -- but, this may cause false positives or very borderline cases, and otherwise it's confusing that tweaking `reportable` also affects this. Either way it gives too much importance to `exceptionalDif`, imo, and with the improved `markable` it also should no longer be needed, and just having these as a report is fine.