ensure player names are replaced in PGN exports
closes lichess-org/api#67

Only the PGN tags have the player name replacements,
as explicitely mentioned in the API documentation.

The idea is that if you're getting JSON,
then you can manipulate it and do the needed replacement yourself.
Whereas PGN is often passed directly to replay software,
hence why lichess offers to modify it at the source.