Add optional estimated start time for Simuls
Add an optional estimated start time for Simuls, in case the host wants
to create a simul some time in advance.

Previously the only way to indicate start time (unless immediate...) was
to write a start time in the Simul description. One problem with such a
solution is that the start time from the text won't be displayed in
users local time zones.

This commit adds a field, which is optional, which can be used to give
an estimate of at what time the host intends to click the Start button.