Fix missing first move sounds on mobile devices
Part of a fix for https://github.com/ornicar/lila/issues/8852 (which also needs https://github.com/ornicar/howler.js/pull/4).

Currently the first time a sound is played on mobile safari it need to have been unlocked by a user interaction. In order for Howler to do this it needs to have attached various touch event listeners to the page, and if does this when creating a new Howl for the sound file, which Lila does when loading a sound for the first time.

Note that `lichess.sound.play` calls `loadStandard` the first time a sound is played, but by then it's too late to be part of the interaction, so the first move sound isn't played. We have to preload to solve this.