Make the interface of renderMaterialDiffs less error-prone
Recently, there have been a few bugs that rendered the material
difference on the wrong side of the board. For example, in #9573 the
material diff was on the wrong side in studies that started from black's
orientation. This was fixed in #9583 but this introduced the same bug
for normal analysis boards.

The interface of renderMaterialDiffs was rather confusing and
error-prone as it had a "flip" argument and it wasn't entirely clear
what this was supposed to do with respect to the "player" and "opponent"
arguments, especially in the case of studies (what is the interpretation
of player and opponent here?).

Since renderMaterialDiffs doesn't care about players and flipping but
only cares about which color is shown on which side of the board, this
commit tries to simplify its interface. Flip, player, and opponent are
all replaced with s single "bottomColor" argument. An extra argument
"showChecks" is added (for three-check) because this information was
previously retrieved from the player/opponent arguments.