fix double like (or unlike)
Fixes https://hackerone.com/reports/1500457 and
https://hackerone.com/reports/1500956.

Users could increase or decrease the denormalized like counter arbitrarily.
Split the query to ensure the like count and rank matches a value that
has really been achieved.

Due to the unsynchronized read-modify-write there is a race condition with very
limited impact that I consider to be acceptable. It was also present before
this patch.