user configuration of forum pagination size
Forum topic and post max per page can now be user selected in browser Preferences->Site Settings.

default values no longer can be set in config, in this branch they are hard coded in lila/pref/Pref.scala with the rest of the default pref values.

This was created as a learning exercise and was not in response to any open issue.  Seems to work tho.