fix for 9572
don't show multiple posts from the same topic in recent posts pane at the bottom of lobby view.

the strategy was increase database objects requested from 10 to 30 (to maximize the chance we get at least 10 unique topics).  Then merge posts into a hashmap keyed on topic name where only the text from the most recent is kept as well its author, but we also track the overall number of new posts and all the new post authors per topic. this is all displayed in the same space in the view pane with a compact representation.  finally, we make the table scrollable because we usually end up with more than 10 rows.