remove PostView.topicLastPage in model.scala
This field is unused and unnecessarily complicated the view functions in PostApi.  If the need for it arises somewhere, maybe it can be calculated there.  This change was originally made in a feature branch, but was broken out into a separate PR where it is more likely to be accepted.