publishing this to github
It strikes me that there's some stuff in PostApi that could be done concurrently by being better with Futures.  In light of that, I decided to make everything even worse with additional synchronous `for {} yield` and `flatMaps`.  (seriously tho, scalability seems more important than minimizing latency on a single request).  As always, pointers are welcome.