fix for 9572
#9572

reverted my previous changes which showed number of posts and recent authors per topic.  now they get the same as before but we discard all but the latest post per topic.  since we fetch 30 results now, might as well use them if we end up with more than 10 topics, set a max-height and overflow-y is auto.  we can always .take(10) if this is a problem.