disable local storage on lichess.org/analysis
one dude actually thinks lichess is pressuring him to make a study with the pane reappearing the way it does.  maybe your first impression was the right one?  if you're firm on keeping temp storage live on tools board, we'll definitely want to flush the synthetic local moves before importing PGN or FEN as that is currently broken.