start tracking last language used in db
For now, it's just displayed by the thumb image.  But we could eventually fire off an auto inbox msg from lichess when a youtuber goes live with "lichess.org" in stream title but no stream language. This could tell the offender they won't be featured till they fix it and link to helpful info.