speed up Future.withTimeout with NoStackTrace
and ensure that all callers provide basic debug info

Calling `withTimeout` instanciates the timeout exception
even if the timeout isn't reached. Exceptions are expensive
because they produce a full stack trace.
NoStackTrace allows to disable it.
In the case of future timeouts, the error message should be
sufficient, we don't need the stacktrace.