Fix (one of) the default setting(s) for playBellSound
I think that default got mangled in a merge somewhere, but this is a PR just to make sure.

I wonder if there's a better way to handle defaults for boolean prefs.  I appreciate the explicit nature of getOrDefault obviating the need for things like (x !== false) but it would be nice to have a single authoritative list of defaults for local prefs.