Expand lang regex in route to take into account Kurdish (Kurmanji)
This add additional complexity on the router side, but I think it's negligible, since those pages are not hit a lot.
The other solution would be to assign the `ku` lang code to `Kurdish (Kurmanji)` instead of `kmr`, since we do not support other kurdish dialects.
This solution would potentially be problematic with crowndin though.