Make rating range unselectable for provisional ratings
Those range are ignored anyway by the pairing system. Greying out is better UX than just removing it I think, as with the tooltip they can learn why it's not possible. I'm not sold on the wording of that tooltip, but could not find better without being too lengthy.

Fix #12271