Show to admins login attempts with weak passwords
See reasoning for the change: https://hq.lichess.ovh/#narrow/stream/8-dev/topic/weak.20password.20support

Currently stored in a size-limited in-heap cache since it's not data we plan to keep, but it could be inconvenient during frequent restarts.

Data could also be retrieved via the dev CLI instead of shown on the profile.