Study chapter: fix PGN to clipboard for Safari
The issue was that we waited for the PGN http request to return before calling the clipboard API, and at that time it was already disabled back by the browser.
Safari fix is a simplified version of https://developer.apple.com/forums/thread/691873. Unfortunately, Firefox does not support the `ClipboardItem` so made a workaround.

Tested on Safari and Firefox.

fix https://github.com/lichess-org/lila/issues/12461