Use `mini-game` for broadcast multi-boards
The snabbdom code was mainly taken from the code in swiss/arena to show ongoing games.

Tested on both finished and ongoing games, as well as update via ws. Should allow "easy" integration in a 2nd time (even though for now mini-game has a limitation about clock, re-putting it at the beginning of the turn at every refresh, that should be fixed for classical)