Fix Copy PGN for chrome
Apparently on Chrome you need to specify the type of the Blob `{ type: 'text/plain' }`. This seem a bug in Chrome since it apparently work on 112.2 but fail on 112.0.56.

I've left the `console.log` since it's useful in the future since it should only be triggered when something on our end is not working as expected.

Tested on Chrome, Safari and Firefox

fix #12737