top-bar challenge/notifications: use <button>
These two toggle buttons used an <a> element despite lacking an href
attribute. This makes screen readers and using tab to navigate the site
skip over these elements since they appear non-interactive. (A link with
href can't be navigated to). To make it more semantically correct, and
indeed picked up by navigating this way, this replaces them with
<button>.

This does require a CSS change (to remove user agent default background
and border), and a small JS changes, so the code looking for a.toggle in
the top bar now looks for button.toggle instead.