JsDump.translatedJs: no quantity if there's only 1
Since
https://github.com/lichess-org/lila/commit/a5731c21f565006e49b4f36b16b00356fa49624c
removed plurals items that had identical content, JsDump translate would
in some cases only add translation items for quantities that aren't
"other", which meant the quantity suffix was included in the key.

If JS code then tried to use this for a quantity that did not match the
single one included, it would not find the translation key and instead
print the translation key where, for example "3 players" should be.

This change detects when there's only a single item in a <plurals>, and
then drops the quantity suffix in the key for the generated JS, ensuring
it will always pick that one.

Thanks to @benediktwerner for helping hunt this one down.