Fix double correspondence email notification
The function `tick()` is run every 10 min, but the emails are sent in a timeframe of 11 minutes, so they can be sent at 5:00 and 5:10. That minute laps was set to be sure at least once per day an email would be sent, but it seems unecessary and produce way too more duplication emails.

fix https://github.com/lichess-org/lila/issues/12688