Ublog: Increase max image height
As long as the width is kept to 100% it's still easy to read/see by scrolling. This notably will allow data-analysis figures to be more readable, as they're usually the pictures with high heights.

For example, after the change https://lichess.org/@/oortcloud_o/blog/tell-me-the-elo/JGcZSVhN is much easier to read (can be tested locally by overriding the value in browser). The `450hv` comes from this blog post.

The introduction of the new variable `$img-max-height` is to minimise disturbance as it is also used in team descriptions or class, where larger images may not be wanted.