Memorize appeal filters
Use a cache to store the filter for each mod, so that it persists across page loads.
I tried to pass the filters by value at first but this was not clean because it is called from too many places.

I don't know if it's the best type of cache, it's not clear what `notLoading` means.

There's also a bit of hack to reset the filters, and to distinguate "I want all appeals" from "I've not set the filter, use the store cache".

I *think* it could be better handled if it was something like `Option[Set[Filter]]` instead of `Option[Filter]`, but could also just add more complexity.

`Option[Set()]` would mean all appeals, while `None` would mean cache. But having empty Set mean all appeals is inconvenient for coding I think.

tested