add a Flying Spaghetti Monster flag
As much as I think lichess should remain agnostic and independant from any religion...

Pastafarianism IS the only true religion.
And there is proof, as it is written in the holy website: https://www.spaghettimonster.org/about/

So personally I think we can make an exception for it.

May we all be touched by His Noodly appendages. R'amen.