Use gregorian calender and western numerics for arabic users
For many users, using the islamic calendar is not practical on the internet, due to international context, so we make sure it's displayed using the gregorian calender.

This was reported by and confirmed by native speakers. ar-ma is a locale which displays the date in arabic while still using gregorian calender.