change 'tactical puzzle' to 'training puzzle'
Puzzles do not have to be tactical and on the growing lichess themes page https://lichess.org/training/themes there are puzzle themes that focus on regaining equality, defensive moves etc. Replacing 'tactical' with 'training' (also the URL branch) allows for various puzzle themes. It is a broader, better term