Send reminder to set email to old account
Only happen when logging-in because that's when we already fetch the email. I have not found a better spot to check it for now, even though some users stay logged in for a very long time.

Only check for the `email` field, because I assumed the `verbatimEmail` is always set in pair with it.

The message is not translated for now.