Fix rating diff being undefined when solving unrated puzzles
Fixes #14393

The issuse was created #14277 which itself was introduced by 4a54a23. More specifically a76f4c3 and 0649bc5.

It was caused by changing to `looseH` which cannot handle a number, and which tried to re-unify 0 and null as the same value, which is not possible in this case. I rolled back to the old version.

I chose to use the original `h` function, so had to revert the `looseH` alias. It's more code but felt cleaner than aliasing h.

Tested for:
- rated
- unrated
- streak