Make position evaluation glyphs available via shortcuts
When using lichess to analyze openings or solve chess studies, one needs to frequently select a position evaluation annotation like `=` (equal) or `+-` (white is winning) – much more frequent than _Zugzwang_ or similar symbols; more times than the existing move annotation glyphs (`?!` `?` `!!` etc.).

Use Shift+1 to Shift+8 for these, analogous to 1 to 6 for the move annotations.

Likely, the help text for move annotations should be changed (no chess player knows what a _glyph annotation_ is, but a _move annotation_ is widely known), but as I am unfamiliar with the i18n system, I'll leave that for a future PR.