Remove restriction on sending the message "You too!"
There was a (I assume accidental) restriction on when someone is allowed to send the message
"You too!" in a game discussion.
Presumably this is so that the button to automatically say this after someone else auto-sends
"Good game!" doesn't work, but this also prevents someone from typing the message manually (so
long as no other person has written a message so far in the chat).