Make DGT board Play Menu shortcut configurable
This commit adds the possibility to configure the inclusion or exclusion
of the "DGT board" play menu entry, via the /dgt/configure page.
Before, the "DGT board" entry would be automatically added to the menu
upon generating a token (`/dgt/config/token`) and/or playing
(`/dgt/play`), and there was no way to remove it.