Mod zone: Fix inquiry button border radius.
Due to nesting a `btn-rack__btn` inside a `form.btn-rack`, the `:last-child` selector of `rack__btn` was triggering as if the button was on the right-side of the rack, despite being on the left side.

The `4px` was chosen because it works well, `5px` was leaving some padding between the two elements.

Properly computing the right radius is not trivial: https://stackoverflow.com/questions/57516318/how-do-i-properly-match-my-child-element-with-its-parents-border-radius