Adapt some implicit resolution change in 3.5
There is a conflict between opaqueHandler and userIdOfWriter for UserId
Both are sastify for BSONWriter[UserId]. In this case of conflicting,
before 3.5 the compiler will pick the more specific type (in this case
opaqueHandler: BSONHandler[UserId] because it's more specific than
BSONWriter[UserId]). In 3.6 they will flip the priority, a.k.a pic the
most general. So, this specific case doesn't really affect us. But
removing implicit ambiguity is good anyway I guess.

More detail on the implicit resolution change here:
https://github.com/scala/scala3/pull/19300