Puzzle: only show feedback on your moves
Consistent feedback requested it, and prior behavior was a bit strange, especially after capture where the checkmark would now be on the piece of the opposite color.

This has the side-effect of effectively disabling the checkmark for users with animation set to "none", and I'd argue it's better and consistent.

See attached videos for the before and after.