broadcast sync: always store delay PGN even if response is cached
if 2 broadcasts share the same source URL,
and only one broadcast has a delay,
then if the non-delay broadcast syncs first,
it populates the dedup cache, which is then
hit by the delayed broadcast.
Therefore we need to store delay PGN in the case
of a cache hit as well.