Restore old tourney scheduling behavior
(temporarily) restore old tourney scheduling logic.

This scheduling is non-deterministic when two tourneys abutt
each-other, but will eventually work since the schedule is tried
frequently.

This commit goes against the long term plan of making the tourney
schedule consistent, clear and deterministic.  The long term
plan is to ignore conflicts caused by the staggered start of a tourney
(probably by using it's planned start time) and to display
tourneys using their unstaggered starts on the tourney page.

So this commit will be reverted, but is needed short term.