Tournament Scheduler: Simplify private "at" method
`at()` does not generally need to return an Option. Reflecting the
intention of the method, we now specify hours/minutes exactly instead
of adjusting a LocalDateTime.

The existing impl returns Some[LocalDateTime] for nearly any hour value,
even though arbitrary values (such as negatives) are not an intended use
case of `at`.