allow challenge API to ignore moreTime rules and preferences
TD needs to adjust time, while TD does not want players to add/adjust
time by themselves. Use case:Armageddon games2.Make correction to
Individual boardCase for no.2. We sent out links, we started the time
too early, TD made mistake. We need to adjust time.Case for no2. Player
said, it was opponent turn, but actually it was his turn. (Bug that I
see sometimes, rare but happened, could be internet problem). TD checked
evidences, and approved it, so we need to adjust time.Case for no.2. One
player violated small regulations, but not enough to make him lose the
game. Usually we take time from his side, but if that is not possible we
can add time to his opponent.The problem with current setting is that
"allow add time" is bind to the player capability and we do not have
setting "player allow add time" and "admin allow add time". Admin in
this case is the account (Auth) that created the board.