broadcast tree sync from multiple contributors and relay source brainfuck
the chapter already has all the game moves, but
its relayPath might be out of sync. This can happen if the broadcast
has contributors who use REC to record and share variations while the
broadcast is ongoing.
If they record a variation that is
then played out by the broadcast players, then there are
no moves to add and send to clients, but the relayPath needs to be updated,
both in the database, and in the clients browsers.
To achieve this without adding a new websocket event type, we send the last
game move again, which contains the relayPath.