wallpaiper over preview option in WYSIWYG editor
It does not show what your blog post will be rendered as and can cause
confusion see the issre this closes. It is a quick and dirty solution to
remove a pointless tab pane that only causes confusion and has no point.

A more involved fix would involve getting the preview and result to line
up. But I am just an intern or something.