API: Sort team arena by ongoing, created, finished
https://lichess.org/api#tag/Teams/operation/apiTeamArena

Before this commit, the results from the endpoint
`/api/team/{teamId}/arena` are found from the database with a very clean
and simple query:

    coll
      .find(forTeamSelect(teamId))
      .sort($sort.desc("startsAt"))
      .cursor[Tournament]()

It adheres to the API documentation,

    "Tournaments are sorted by reverse chronological order of start date
    (last starting first)."

```
Status
10 - Created
20 - Ongoing
30 - Finished
```

```json
{ "startsAt": "[99 days left]", "status": 10, ... }
{ "startsAt": "[98 days left]", "status": 10, ... }
...
{ "startsAt": "[1 hours left]", "status": 10, ... }
{ "startsAt": "[0 hours left]", "status": 20, ... }
{ "startsAt": "[1 hours past]", "status": 30, ... }
{ "startsAt": "[3 years past]", "status": 30, ... }
```

After this commit, the results are sorted by both state and nearest
start time.

The query is now a beast, an aggregation with a "switch" operation, for
sorting on a synthetic status field instead of the original status
field.

https://www.mongodb.com/docs/manual/reference/operator/aggregation/switch/

    "Tournaments are sorted by states ongoing, upcoming and finally
    finished, and a chronological order of start date (first starting
    first)."

```json
{ "startsAt": "[0 hours left]", "status": 20, ... }
{ "startsAt": "[1 hours left]", "status": 10, ... }
...
{ "startsAt": "[98 days left]", "status": 10, ... }
{ "startsAt": "[99 days left]", "status": 10, ... }
{ "startsAt": "[1 hours past]", "status": 30, ... }
{ "startsAt": "[3 years past]", "status": 30, ... }
```

why?

- This commit is based on the bold claim that an API consumer will be
  most interested in ongoing tournaments, and then any upcoming
  tournaments, in the order of the nearest upcoming first, and lastly
  interested in already finished tournaments, by so much certainty that
  the default behaviour should be changed(!)

The use case (before this commit) which is kind of awkward, is for API
consumers who are interested in the nearest ongoing or upcoming
tournaments. They will need to sift through "all" the arenas until they
find a finished arena - to know that the last non-finished arenas are
the nearest.
Maybe the "all" number is typically small, and is what makes the very
clean and simple query so sweet.
But if the "all" number is big, bigger than 100 (default number of
arenas to fetch if `max`-parameter isn't specified), the awkwardness
hits! One will need to guess the amount of "all", so that the guess
includes at least 1 finished tournament to be able to know that the
nearest tournaments have been seen.

Why not?

- Changing a default behaviour is not nice towards existing consumers
- It might be rare that teams have many pre-created tournaments,
  i.e this doesn't need to be changed
- This can be solved in other ways, like implementing more parameters in
  the API request (filter on status, etc)
- The increased complexity by going from a very clean and simple query
  to a complex beast of a query is just unacceptable
- The commit author doesn't know anything about mongo db queries (but
  surely appreciated the opportunity to get some experience with them!)

Example of team with many tournaments,
curl "https://lichess.org/api/team/--elite-chess-players-union--/arena?max=1"