ui/build/sass: Fix theme color fallback when generating color mixes
With `capturedColors` outside the loop, colors that aren't
defined for a theme were taken from any previously loaded
theme where they were present, instead of always falling
back to the default theme.

Specifically, this caused the transparent theme
to incorrectly fall back to the light theme, among other
things leading to bad contrast when the background is a
color mix (from the light theme) but the text color is
not (and therefore from the dark theme).