hotfix changing piece theme on broadcast tour page - type issue
that page doesn't have a main board
https://lichess.org/broadcast/grenke-chess-festival-2025--freestyle-chess-open/round-2/AprGXQxU

therefore ui/analyse/ctrl doesn't always have this.chessground and is
wrongly typed.

however the asumption that there is a board is quite useful and almost
always true, so maybe it should have a headless chessground instead of
undefined when there's no main board?