Remove dangling `SaveStudy` for Practice, and confirm lack of cache invalidation for practice structure
The practice studies are defined as a json-like structure. Every time a study is saved, it should be checking in that study is in practice list, and invalidate the practice cache if needed.

This has been not working for more than a year, since https://github.com/lichess-org/lila/pull/15877, and more precisely 6f6372aad912ce9592a190dcb17d96649ad25fea.

Considering how seldomly practice config is changed and that the cache is 3h, it seems fine not to have invalidation, so remove the now-unused parts left.