probably solve puzzle db performance issues by solving batch?nb=1 selections
both apps can request puzzles in batches, to keep in offline memory.
therefore they use a separate selection mechanism that doesn't store
sessions.

But in practice they almost always request batches of 1 puzzle:
https://monitor.lichess.ovh/d/d36KC5xMz/puzzle-selector?orgId=1&from=now-30d&to=now&timezone=utc&var-instance=lila@ocean&var-field=mean&viewPanel=panel-31

and they do so after each puzzle is played.

as a result to we don't get benefits of batching,
and we don't get benefits of storing user puzzle sessions either.

after this patch, the batch selector branches to the single puzzle
session selector in the case of `?nb=1`.

I think it should help database performance a lot.