always generate a correct manifest in watch mode
if the noCheck phase of tsc took longer than the noEmit one, which can sometimes happen when changing something small like a comment, it overwrote the successful noCheck status with undefined because it mistakenly assumed the check will come in afterwards. this meant that a dirty manifest check never triggered the manifest write. meanwhile, esbuild cleans up after itself, so it deleted the original bundle that contains the change when it created the new one with a different hash, rendering the current (stale) manifest invalid and generating page load exceptions.