optimize fide_player_follower
we can skip the `p` field and more importantly the index it would
normally require,
because a regex like `_id: /^1503014\//` uses the _id index efficiently.

https://www.mongodb.com/docs/manual/reference/operator/query/regex/#index-use

We can also use `distinct("_id", ` so that only the index is used
without ever fetching the document.

We still want the `u` field so we can count and list follows by user.