Fix for Tournaments with no streak bonuses should still have wins coloured green #17880
When streaks are disabled, the renderScoreString wraps each score in a <score> tag with no further differentiation, meaning all wins are rendered with the neutral <score> tag (which appears grey by default) instead of the green <streak> tag.

Instead of returning all scores as <score> for non-streak tournaments, we return wins with the <streak> tag.

The change made wraps any score greater than 1 (i.e. 2 points for a win, or 3 points for a berserk win) in a <streak> tag even when streaks are disabled, while leaving draws and losses as <score>. This means every win will be colored green, as desired, regardless of the streak setting.