remove green "create game" button outline from setup dlg in desktop views
it looks neat. unfortunately, it also looks like a focus ring when that element has no focus. 

this is a PR rather than an issue as the bare button seems fine to me, but maybe you can find a way to style things so it still pops without suggesting keyboard focus. the outline remains on touch devices.