more approachable markdown reference
the [github reference](https://docs.github.com/en/get-started/writing-on-github/getting-started-with-writing-and-formatting-on-github/basic-writing-and-formatting-syntax) is like diving into the middle of a technical manual. from pretty early on, it shifts focus to various github features and things that have nothing to do with markdown.

this [cheat sheet](https://www.markdownguide.org/cheat-sheet/) is more approachable and elements covered are a better match for what lila's flexmark config offers.

markdownguide.org materials are open source from https://github.com/mattcone/markdown-guide. he does advertise a paperback of his guide at the bottom of every page.