minor relay sync refactor, remove comment where the code is expressive enough
in my opinion the code is readable enough, that a comment explaining
what it does is not necessary.

I like to comment only obscure code, when the intent is mysterious. And
even then a commit message is usually preferable.

Comments don't compile, they age and become outdated: they have a
maintenance cost.

Commit messages are never outdated because each change brings a new one.