Theme-vote API: fix `checkFirst` theme read from db
It was the root of the issue causing 400 malformated requests for all puzzles containing the `checkFirst` theme, like `/training/02tYt`.

It was because it would fail mapping the BSONDocument/Bdoc to `PuzzleRound.Theme`.

```scala
PuzzleTheme
        .find(v.tail)
```
in the following snippet would fail because internally it's only looking at `visible` puzzles, which `checkFirst` is not.

```scala
  private[puzzle] given BSONHandler[PuzzleRound.Theme] = tryHandler[PuzzleRound.Theme](
    { case BSONString(v) =>
      PuzzleTheme
        .find(v.tail)
        .fold[Try[PuzzleRound.Theme]](handlerBadValue(s"Invalid puzzle round theme $v")) { theme =>
          Success(PuzzleRound.Theme(theme.key, v.head == '+'))
        }
    },
    rt => BSONString(s"${if rt.vote then "+" else "-"}${rt.theme}")
  )
```
so would fail, resulting in the following failing

```scala
    def themes(id: PuzzleRound.Id): Fu[Option[List[PuzzleRound.Theme]]] =
      colls.round:
        _.primitiveOne[List[PuzzleRound.Theme]]($id(id), PuzzleRound.BSONFields.themes)

```
resulting in returning `None` instead of a `List[PuzzleRound.Theme]`.

I was surprised by that behavior, I would have expected the parser to just skip the malformated elements, but still produce a List.

I guess it's the difference between `Option` and `Try` for parsers.